git-svn-id: http://typo3scripts.googlecode.com/svn/trunk@35 a8a2147e-0eb0-e98b-0563-54e275c8f842

+++ createBackup.sh
@@ -4,6 +4,9 @@
 # written by Oliver Salzburg
 #
 # Changelog:
+# 1.4.0 - Code cleaned up
+#         Extended command line paramter support
+#         Improved self-updating
 # 1.3.4 - Fixed update location
 # 1.3.3 - Now using generic config file sourcing approach
 # 1.3.2 - Now using explicit modifiers
@@ -58,15 +61,42 @@ UPDATE_BASE=http://typo3scripts.googlecode.com/svn/trunk
 # Self-update
 runSelfUpdate() {
   echo "Performing self-update..."
-  wget --quiet --output-document=$0.tmp $UPDATE_BASE/$SELF
-  chmod u+x $0.tmp
-  mv $0.tmp $0
-  exit 0
+  
+  # Download new version
+  echo -n "Downloading latest version..."
+  if ! wget --quiet --output-document="$0.tmp" $UPDATE_BASE/$SELF ; then
+    echo "Failed: Error while trying to wget new version!"
+    echo "File requested: $UPDATE_BASE/$SELF"
+    exit 1
+  fi
+  echo "Done."
+  
+  # Copy over modes from old version
+  OCTAL_MODE=$(stat -c '%a' $SELF)
+  if ! chmod $OCTAL_MODE "$0.tmp" ; then
+    echo "Failed: Error while trying to set mode on $0.tmp."
+    exit 1
+  fi
+  
+  # Spawn update script
+  cat > updateScript.sh << EOF
+#!/bin/bash
+# Overwrite old file with new
+if mv "$0.tmp" "$0"; then
+  echo "Done. Update complete."
+  rm \$0
+else
+  echo "Failed!"
+fi
+EOF
+  
+  echo -n "Inserting update process..."
+  exec /bin/bash updateScript.sh
 }
 
 # Read external configuration
 CONFIG_FILENAME=${SELF:0:${#SELF}-3}.conf
-if [ -e "$CONFIG_FILENAME" ]; then
+if [[ -e "$CONFIG_FILENAME" ]]; then
   echo -n "Sourcing script configuration from $CONFIG_FILENAME..."
   source $CONFIG_FILENAME
   echo "Done."
@@ -112,14 +142,39 @@ fi
 
 # Begin main operation
 
+# Does the base directory exist?
+if [ ! -d $BASE ]; then
+  echo "The base directory '$BASE' does not seem to exist!"
+  exit 1
+fi
+# Is the base directory readable?
+if [ ! -r $BASE ]; then
+  echo "The base directory '$BASE' is not readable!"
+  exit 1
+fi
+
+# Filename for snapshot
 FILE=$BASE-$(date +%Y-%m-%d-%H-%M).tgz
+
 echo "Creating Typo3 backup '$FILE'..."
+
+# Create database dump
 echo -n "Creating database dump at $BASE/database.sql..."
-mysqldump --host=$HOST --user=$USER --password=$PASS --add-drop-table --add-drop-database --databases $DB > $BASE/database.sql
+if ! mysqldump --host=$HOST --user=$USER --password=$PASS --add-drop-table --add-drop-database --databases $DB > $BASE/database.sql; then
+  echo "Failed!"
+  exit 1
+fi
 echo "Done."
+
+# Create backup archive
 echo -n "Compressing Typo3 installation..."
-tar --create --gzip --file $FILE $BASE > /dev/null
+if ! tar --create --gzip --file $FILE $BASE > /dev/null; then
+  echo "Failed!"
+  exit 1
+fi
 echo "Done."
+
+# Now that the database dump is packed up, delete it
 echo -n "Deleting database dump..."
 rm --force $BASE/database.sql
 echo "Done!"
